AIA-Noobs/requirements.txt
monai==0.9.1
numpy==1.23.1
opencv-python==4.5.5.64
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
scipy==1.9.0
six==1.16.0
torch==1.12.1
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
PyYAML==5.4.1
imgaug
segmentation-models-pytorch
pytorch-ignite
wandb
git+https://github.com/lessw2020/Ranger21.git